KiloInfinite Application Setup, Kilo MVC Base Established
Application setup and running in Localhost, Kilo model, controller and
views have been setup and are working efficiently.

Kilo records are saving to the database when formed via the Rails console,
and are then displaying in the show page when the specific record is called.

Next code published will be to the README.