Devise and Users Setup
Following the previous PR adding Sidekiq and routes for user authentication,
this made a requirement for users and authentication for them to be established.

Latest version of the Devise gem added to the Gemfile and installed.

Added Devise users model to the application via the rails generate devise User command.

Users can now be created via the user sign up page and exist when searched for in the
Rails console.

The end goal is to incorporate user mailers, Sidekiq and users together with this continuation
from Sidekiq into users and user authentication, further user code will be implemented
followed by user mailers.