Navbar Added via a Partial, Devise Views Implemented
Added the latest version of the Bootstrap gem along with Dart Sass Sprockets
to handle the Sass engine required for Bootstrap.

Added Devise views via the standard generate command to incorporate Devise
further into the application.

Shared flashes partial created and code added to it to allow flashes to display
for alerts and notices.

Partials for the navbar and flashes have been added to the application layout file
to appear across the application.

Application controller updated to require user authentication before they can perform
CRUD actions, while in the kilos controller access to the index page is granted to
users without requiring a login.

Converted the application.css file to .scss to enable Bootstrap funcitonality and
customisation going forward.

Added a Bootstrap navbar to the navbar partial created in
the shared views folder.