Added Weight to Kilo
Weight now exists with the Kilo architecture, as a numerical value that
needs to be passed into a Kilo record now to create a record.

Added weight as an input in the new view and form partials. Parameter
added to the private method in the kilos_controller also.

Confirmed that weights will appear from records created both in the Rails
console and the localhost server.

Updated commented instruction code in the kilos_controller.