pg_search gem installed and setup in the kilo model
Latest version of the pg_search gem has been setup to handle
search functionality within the application.

Implemented pg_search multisearch so that "one global index is created for all the active record classes"
according to https://www.mintbit.com/blog/searching-made-easy-with-pg-search-gem-in-ruby-on-rails.

Records can be searched in the console via the PgSearch.multisearch("example").