Infinite Controller, Views, Migration and Seeds Added
Added and performed a migration for the name attribute and its
validation for the Infinite model.

Basic Infinite controller code added for several CRUD methods.

Show and index view pages setup for Infinite records, implemented
same Turbo code for Infinite index records as in the Kilo index.

Setup seeds to create a total of 20 Infinite records, 10 of Wine, and
10 of Milk.