Infinite Form Partial and Relevant Controller Code Setup
Used Simple Form again in the same style as in the Kilo form partial
to create an Infinite form partial. The partial sits within the
Infinite index page also replicating the Kilo placement of the form.

I have also updated the infinites_controller to pass the infinite
parameters into a private method that are then called in the controller
create action. Along with adding the "new" infinite method line to the
index controller action to faclitate the inclusion of an Infinite form
in the index view.