Modal with Popover inside the Modal Added to Kilos Index Page
A Bootstrap modal with a popover inside the opened modal has been
added to demonstrate the depth of which a single webpage can be
customised without the browser reloading or redirecting to a different
webpage.

This Bootstrap modal with a openable popover exists within the
Kilos index page/current application homepage.