Images Added to the Assets, one Image Displayed in Frontend
Added several images to the application assets folder, and a .keep
file for the new folders created in the images directory: kilo and
infinite.

Kilo index view features the new dumbbell.jpg image and a smaller resized
calling of the image as it was too large for the page in its original form.

Updated the README document with URL's for where the images have been
sourced from.