Action Policy Gem Added to the Application
Authorisation gem Action Policy has been added to the application to control user activity.

The latest version of the action_policy gem has been installed (0.6.9).

A Kilo policy file has been generated and added so that all users are
allowed to perform "index" activities for the Kilo section of the application.