RSpec and Factory Bot Setup
Latest versions of RSpec(6.1.2) and Factory Bot(6.4.3) gems installed.

rspec_rails has been installed and added within the group :development, :test
section of the Gemfile, and the same for the factory_bot_rails gem.

Factory Bot has been added to the requires in the spec/rails_helper/rb file.

I have added a factories folder for factory bot and .keep file to maintain
the existence of the factories folder until files are added within it.