Improved Log In and Sign Up Pages Appearance, Implemented Log Out Functionality
Added a new CSS file for elements within updated Devise sign up and log in pages
to make them appear closer to the design for the rest of the website.

Also added a log out submit link to the navbar dropdown, and updated the
Devise initializer file from a delete to get method to allow the sign out
to occur (otherwise it will result in a Rails error).