Deletion of Records now possible within Kilo and Infinite
Turbo deletion actions have been added to the frontend show pages for
individual kilo and infinite records. Once a record is clicked on and
deleted the loggged in user (login is required for this) is redirected
to the respective index page (either kilo or infinite depending on
whether a kilo or infinite record was deleted).

The kilo and infinite controllers have also been updated to allow the
deletions to occur, with both controllers having a method to find the
specific record the user wishes to delete before then deleting it.