New Engima Model and Controller
Looking to bring in more abstract Ruby and Rails into the application
without interfering too heavily with the existing Kilo and Infinite MVC
architecture, introducing Enigma is the chosen way for doing this.

Created an Enigma model with paradox and body attributes to it (will likely
add further attributes soon) and validation for these attirbutes also set
in the model.

Migration file created and migrated into the DB schema for the new Enigma
model.

Also generated the base Enigma controller to work on further soon.