User Expanded with First and Last Name
The user model and Devise sign up view have been expanded with the inclusion
of first and last name fields, along with validations inserted for the already
existing email and password fields, to ensure they are not extremely long in
length and likely not an email/password.

A method for the new inclusion of first and last name for a user has been
setup in the user model as "name".

Migration files have been created for both adding a name to a uuser and first
and last name to a user, these have both been independently of each other
before being migrated into the DB schema.